Refactor project structure: reorganize sections and components, enhance styling for AboutMe, Projects, and Footer sections, and implement responsive design. Remove unused files and update CSS for better layout and alignment.
