+++ src/App.jsx
@@ -1,12 +1,12 @@
 import "./index.css";
 import "./Components/Card.css";
-import { AboutMe } from "./Sections/AboutMe.jsx";
+import { AboutMe } from "./Sections/MainBody/About/AboutMe.jsx";
 import { Header } from "./Sections/Head.jsx";
-import { Projects } from "./Sections/Projects.jsx";
+import { Projects } from "./Sections/MainBody/Projects/Projects.jsx";
 import { Toolbox } from "./Sections/Toolbox.jsx";
 import { Skills } from "./Sections/Skills.jsx";
-import { Updates } from "./Sections/Updates.jsx";
-import { Footer } from "./Sections/FooterContent.jsx";
+//import { Updates } from "./Sections/Updates.jsx";
+import { Footer } from "./Sections/Footer/FooterContent.jsx";
 
 export const App = () => {
   console.log("App.jsx");
@@ -20,7 +20,6 @@ export const App = () => {
         <Skills />
         <Projects />
         <Toolbox />
-        <Updates />
       </main>
       <footer>
         <Footer />
