+++ src/Components/ContactDetails.jsx
@@ -4,9 +4,15 @@ export const Contacts = () => {
   return (
     <div className="contacts">
       <ul>
-        <li>Jonny Hicks </li>
-        <li>+46 76 085 08 54 </li>
-        <li>hello.jonnyhicks@gmail.com</li>
+        <li>
+          <h4>Jonny Hicks </h4>
+        </li>
+        <li>
+          <h4>+46 76 085 08 54 </h4>
+        </li>
+        <li>
+          <h4>hello.jonnyhicks@gmail.com</h4>
+        </li>
       </ul>
       <ButtonLinks />
     </div>
