+++ src/Components/NetlifyButton.css
@@ -1,5 +1,5 @@
 .netlify-button {
-  border-radius: 12px;
+  border-radius: 1rem;
   box-shadow: var(--shadow);
   cursor: pointer;
   font-family: "IBM Plex Mono", sans-serif;
@@ -18,4 +18,7 @@
 .netlify-button:active {
   border: transparent;
   box-shadow: inset var(--shadow);
+}
+.netlify-button > li {
+  list-style-type: none;
 }
