--- src/Sections/FooterContent.css
@@ -1,27 +1,0 @@
-.footer {
-  display: grid;
-  grid-template-columns: repeat(6, 1fr);
-  grid-template-rows: repeat(3, 1fr);
-}
-
-.footer-background {
-  grid-column: 1 / span 3;
-  grid-row: 1 / span 2;
-  color: var(--color-primary);
-  background: var(--color-secondary);
-  height: 100vh;
-  width: 100vw;
-}
-.contact {
-  grid-column: 4 / span 3;
-  grid-row: 3 / span 1;
-}
-
-.contact > ul {
-  list-style: none;
-  display: flex;
-  flex-direction: column;
-  align-items: start;
-  padding: 0;
-  margin: 0;
-}
