+++ src/Sections/MainBody/About/AboutMe.css
@@ -1,0 +1,27 @@
+#about-me {
+  display: flex;
+  flex-direction: row;
+
+  padding: 5rem;
+  justify-content: center;
+  box-sizing: border-box;
+  align-items: center;
+}
+.about-text {
+  text-wrap: wrap;
+  align-items: start;
+  justify-content: start;
+  text-align: left;
+  padding: 1rem;
+  margin: 0.5rem;
+  max-width: 50%;
+}
+.about-text > p {
+  text-wrap: wrap;
+  color: var(--font-color-light);
+  font-size: 1rem;
+}
+
+.portrait {
+  width: 50;
+}
