+++ src/Sections/MainBody/Projects/Projects.css
@@ -1,0 +1,25 @@
+.projects {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: start;
+  justify-content: space-evenly;
+  flex-wrap: wrap;
+  padding: 1rem;
+}
+.card {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  flex-wrap: wrap;
+  width: 100%;
+  max-width: 20rem;
+  height: 100%;
+  margin: 0.5rem;
+  padding: 1rem;
+  border-radius: 1rem;
+  color: var(--font-color-primary);
+  background-color: transparent;
+
+  text-align: center;
+}
